+++ packages/go_router_builder/test_inputs/typed_query_parameter.dart
@@ -10,7 +10,7 @@ mixin $OverriddenParameterNameRoute {}
 class OverriddenParameterNameRoute extends GoRouteData
     with $OverriddenParameterNameRoute {
   OverriddenParameterNameRoute({
-    @TypedGoRouteParameter(name: 'parameterNameOverride') this.withAnnotation,
+    @TypedQueryParameter(name: 'parameterNameOverride') this.withAnnotation,
   });
   final int? withAnnotation;
 }